create cudnn_sdpa
[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

typo

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

revert function import

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

spda --> sdpa

fix layernorm input

cudnnex

change the way cudnn is checked

fix import

bind functions

add function def as none

add accidentally deleted code

bind import

[pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
